vector macro used to remove Tiles structure : Tiles -> vec(Tile *)

+++ model/player.c
@@ -8,20 +8,20 @@ Vector2 player_pos(Player i) {
     Vector2 pos;
     switch (i) {
     case Player0:
-        pos.x = WIDTH / 2 - 7 * TILE_WIDTH;
+        pos.x = (float)WIDTH / 2 - 7 * TILE_WIDTH;
         pos.y = HEIGHT - TILE_HEIGHT;
         break;
     case Player1:
         pos.x = WIDTH - TILE_HEIGHT;
-        pos.y = HEIGHT / 2 + 7 * TILE_WIDTH;
+        pos.y = (float)HEIGHT / 2 + 7 * TILE_WIDTH;
         break;
     case Player2:
-        pos.x = WIDTH / 2 - 7 * TILE_WIDTH;
+        pos.x = (float)WIDTH / 2 - 7 * TILE_WIDTH;
         pos.y = TILE_HEIGHT;
         break;
     case Player3:
         pos.x = TILE_HEIGHT;
-        pos.y = HEIGHT / 2 - 7 * TILE_WIDTH;
+        pos.y = (float)HEIGHT / 2 - 7 * TILE_WIDTH;
         break;
     }
     return pos;
